Ordering/main/requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.7
et-xmlfile==1.1.0
h11==0.14.0
idna==3.4
keyboard==0.13.5
openpyxl==3.1.2
outcome==1.2.0
pdfminer.six==20221105
pycparser==2.21
pynput==1.7.6
PyPDF2==3.0.1
PySocks==1.7.1
python-dotenv==1.0.0
pywin32==306
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.22.2
trio-websocket==0.11.1
undetected-chromedriver==3.5.4
urllib3==2.0.5
websockets==12.0
wsproto==1.2.0
xlrd==2.0.1
